Automatic selection for regions with one scale
This implementation of selectable scale types includes various bugfixes, and adds automatic selection of scale for regions with only one scale type. This IS the case for all regions after migration, as they are all assigned a single, default scale- a Bathroom Scale that measures in pounds. Regional admins can modify, add, and delete scales; no matter which scales are present, default or not, a selection field will be rendered unless there is EXACTLY one scale for a region. An empty scale list will result in an empty selection field and an unsaveable log.

Instructions for region admins using this new update:

*   If your region measures deliveries uniformly using a scale that measures in pounds,
        you don't need to do anything. Volunteers can continue using the Robot as normal
        with no action required on your part.

*   If your region measures deliveries with different scales, you can now use the "Scale Types"
        interface located under Regional Admin on your control bar. Because the overall
        Robot uses pounds as a weight measurement unit for statistical purposes, creating a
        scale with a different unit of measurement will result in automatic conversion occuring
        when volunteers save their logs for the first time.

*   If your region has no scales: consider editing the default scale to have a name signifying
        that your shifts are not being measured by an actual scale - this name will be associated
        with you region's logs in the database, which is helpful for statistical analysis.