Add spec for FoodRobot::generate_log_entries
The goal here is to be able to refactor the #generate_log_entries
method. First step is to get a test case around everything so we don't
lose functionality or break things.

The test is slow, but hopefully we can remove it in favor of smaller,
faster tests once the function is refactored.