Extract `Shift` class to encapsulate logic around shifts
A Shift is a collection of logs that belong to the same schedule chain
and occur on the same day.

Logic for building Shifts from a collection of logs comes from
Log::group_by_schedule.

Eventually, the Shift model could (and probably should) be backed by a
SQL table, but for now it will just exist in memory.