Only save associated log parts if they already exist in the database
Fixes #74.

There was an issue where duplicate log_volunteer records were created
when saving a Log with associated volunteers and log_parts created in
memory. The problem only occurred in regions with exactly one scale type
because the newly created logs are assigned that scale type.

We don't want to save the associated records individually as they depend
on the Log record. We just set the attributes on them and let
ActiveRecord save them itself.