Fix flaky ability test
The test has failed in the following circle ci builds:

- https://circleci.com/gh/boulder-food-rescue/food-rescue-robot/164
- https://circleci.com/gh/boulder-food-rescue/food-rescue-robot/180
- https://circleci.com/gh/boulder-food-rescue/food-rescue-robot/185

My hypothesis is that this flakiness is caused by the behavior of the
logs factory when a region is not explicitly provided. The factory picks
a random region if at least 5 exist. That could occassionally pick the
region that the volunteer is actually assigned to, making the test fail.

By assigning an explicitly different region to the log, my hope is that
the test will no longer fail.